+++ Genius/Extensions/Swift/String.swift
@@ -3,9 +3,19 @@
 // See LICENSE.txt for license information.
 //
 
-extension String: DataInitializable, UIStringRepresentable {
+import AppKit
+import SwiftUI
+
+extension String: DataInitializable, UIStringRepresentable, FormValue {
 
 	var uiRepresentation: String? {
 		self
+	}
+
+	var formView: some View {
+		Button(self) {
+			NSPasteboard.set(self)
+		}
+		.buttonStyle(.borderless)
 	}
 }