+++ Genius/Models/Maintenance/MaintenanceCheck.swift
@@ -11,13 +11,31 @@ import SwiftUI
 struct MaintenanceCheck<
 	Value: PossiblyOptional & Sendable,
 	ValueWrapper: ValueWrapperProtocol<Value>,
->: SystemInformationProtocol, UISymbolRepresentable where Value.Wrapped: Comparable & Maximizable & Sendable {
+>: SystemInformationProtocol where Value.Wrapped: Comparable & Maximizable & Sendable {
 
 	typealias Wrapped = Value.Wrapped
 
 	let valueWrapper: ValueWrapper
 	let requirement: Wrapped
 	let available: Bool?
+
+	var syncUIRepresentation: Symbol?? {
+		if !?available ?? false {
+			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : .none
+		} else if let syncValue {
+			if let value = syncValue.optional {
+				value >= requirement ? // swiftlint:disable:this void_function_in_ternary
+					Symbol(.checkmark, color: .green, label: .enabled) :
+					Symbol(.xmark, color: .red, label: .disabled)
+			} else {
+				Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ?
+					Symbol(.questionmark, color: .red, label: .unknown) :
+					.none
+			}
+		} else {
+			.some(nil)
+		}
+	}
 
 	var uiRepresentation: Symbol? { get async {
 		if !?available ?? false {