+++ Genius/Models/SystemInformation/SystemInformationData.swift
@@ -9,10 +9,24 @@ import Foundation
 struct SystemInformationData<
 	Value: UIStringRepresentable,
 	ValueWrapper: ValueWrapperProtocol<Value>,
->: SystemInformationProtocol, UIStringRepresentable {
+>: SystemInformationProtocol {
 
 	let valueWrapper: ValueWrapper
 	let available: Bool?
+
+	var syncUIRepresentation: String?? {
+		if !?available ?? false {
+			Defaults[.developmentMode] ? String(localized: .notAvailable) : .none
+		} else if let syncValue {
+			if let uiRepresentation = syncValue.uiRepresentation {
+				uiRepresentation
+			} else {
+				Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ? String(localized: .unknown) : .none
+			}
+		} else {
+			.some(nil)
+		}
+	}
 
 	var uiRepresentation: String? { get async {
 		if !?available ?? false {