+++ Genius/Models/SystemInformation/SystemInformationProtocol.swift
@@ -3,13 +3,18 @@
 // See LICENSE.txt for license information.
 //
 
-protocol SystemInformationProtocol {
+import _Concurrency
+
+protocol SystemInformationProtocol<Representation>: Sendable {
 
 	associatedtype Value
 	associatedtype ValueWrapper: ValueWrapperProtocol<Value>
+	associatedtype Representation
 
 	var valueWrapper: ValueWrapper { get }
 	var available: Bool? { get }
+	@MainActor var syncUIRepresentation: Representation?? { get }
+	@MainActor var uiRepresentation: Representation? { get async }
 }
 
 extension SystemInformationProtocol {
@@ -17,6 +22,10 @@ extension SystemInformationProtocol {
 	var value: Value { get async {
 		await valueWrapper.wrappedValue
 	} }
+
+	@MainActor var syncValue: Value? {
+		valueWrapper.syncWrappedValue
+	}
 }
 
 extension SystemInformationProtocol where ValueWrapper == SyncValueWrapper<Value> {