+++ Genius/Views/Content/ContentView.swift
@@ -10,13 +10,17 @@ struct ContentView: View {
 
 	@Default(.interfaceMode)
 	var interfaceMode: Settings.InterfaceMode
+	@Default(.developmentMode)
+	var developmentMode: Bool
 
 	var body: some View {
 		AdaptingTabView<ContentViewTab>()
 			.frame(minWidth: 712, minHeight: 256)
-			.uninstallationDialog()
-			.updateInProgressAlert()
+			.modifier(RelaunchDialog())
+			.modifier(UninstallationDialog())
+			.modifier(UpdateInProgressAlert())
 			.id(interfaceMode)
+			.id(developmentMode)
 	}
 }