+++ CHANGELOG.md
@@ -15,9 +15,36 @@ See LICENSE.txt for license information.
 
 ### Security
 
+- Require signed appcast feed
+
 ### Dependencies
 
-### [Commit History](https://github.com/F1248/Genius/compare/v0.4.0...HEAD)
+### [Commit History](https://github.com/F1248/Genius/compare/v1.0.0...HEAD)
+
+## [1.0.0](https://github.com/F1248/Genius/releases/tag/v1.0.0)
+
+### Added
+
+- Add setting to disable Liquid Glass
+- Add setting to hide icons in menu bar
+
+### Improved
+
+- Remove logs when uninstalling Genius
+- Show `ProgressView` in `FormView` only the first time
+- Show `ProgressView` in `FormView` for each entry individually
+- Simplify settings labels
+- Correct English settings labels capitalization
+
+### Fixed
+
+- Fix uninstallation in recoveryOS
+
+### Dependencies
+
+- Update [Xcode](https://developer.apple.com/xcode) from [26.1.1](https://developer.apple.com/documentation/xcode-release-notes/xcode-26_1-release-notes) to [26.2.0](https://developer.apple.com/documentation/xcode-release-notes/xcode-26_2-release-notes)
+
+### [Commit History](https://github.com/F1248/Genius/compare/v0.4.0...v1.0.0)
 
 ## [0.4.0](https://github.com/F1248/Genius/releases/tag/v0.4.0)